# StoryPython/swagger.yml
swagger: "2.0"
info:
  description: This is the swagger file that goes with the API Server
  version: "1.0.0"
  title: USERSTORIES API Server
consumes:
  - application/json
produces:
  - application/json

basePath: /api

# Paths supported by the server application
paths:
  /stories:
    get:
      operationId: stories.read_all
      tags:
        - stories
      summary: Read the entire list of stories
      description: Read the list of stories
      parameters:
        - name: limit
          in: query
          type: integer
          required: false
          description: integer of stories to return
      responses:
        200:
          description: Successfully read stories list operation
          schema:
            type: array
            items:
              properties:
                id:
                  type: integer
                title:
                  type: string
                author:
                  type: string
                createdDate:
                  type: string
                summary:
                  type: string
                tags:
                  type: string
    post:
      operationId: stories.create
      tags:
        - stories
      summary: Create a story and add it to the DB
      description: Create a new story in the Database
      parameters:
        - name: story
          in: body
          description: Story to add to DB
          required: True
          schema:
            type: object
            properties:
              title:
                type: string
              author:
                type: string
              summary:
                type: string
              tags:
                type: string
      responses:
        200:
          description: Successfully created stories list operation
          schema:
            type: object
            properties:
              id:
                type: integer
              title:
                type: string
              author:
                type: string
              createdDate:
                type: string
              summary:
                type: string
              tags:
                type: string
  /stories/{id}:
    get:
      operationId: stories.read_one
      tags:
        - stories
      summary: Read one story from the stories DB
      description: Read one story from the stories DB
      parameters:
        - name: id
          in: path
          description: id of the story to retrieve
          type: integer
          required: True
      responses:
        200:
          description: Successfully created stories list operation
          schema:
            type: object
            properties:
              id:
                type: integer
              title:
                type: string
              author:
                type: string
              createdDate:
                type: string
              summary:
                type: string
              tags:
                type: string
                
  /storytext:
    post:
      operationId: storytext.create
      tags:
        - storytext
      summary: Add the story text to the DB
      description: Add the story text to the DB
      parameters:
        - name: storytext
          in: body
          description: Storytext to add to DB
          required: True
          schema:
            type: object
            properties:
              id:
                type: integer
              story_text:
                type: string
      responses:
        200:
          description: Successfully created stories list operation
          schema:
            type: object
            properties:
              id:
                type: integer
              story_text:
                type: string
  /storytext/{id}:
    get:
      operationId: storytext.read_one
      tags:
        - storytext
      summary: Retrieve one of the stories text from the DB
      description: Retrieve one of the stories text from the DB
      parameters:
        - name: id
          in: path
          description: id of the story text to retrieve
          type: integer
          required: True
      responses:
        200:
          description: Retrieved the story text from the DB
          schema:
            type: object
            properties:
              id:
                type: integer
              story_text:
                type: string